EXLM-2951 Updated Rec marquee and Rec content block coveo value for analytics
