+++ blocks/recommended-content/recommended-content.js
@@ -14,6 +14,7 @@ import ResponsiveList from '../../scripts/responsive-list/responsive-list.js';
 import defaultAdobeTargetClient from '../../scripts/adobe-target/adobe-target.js';
 import BrowseCardsTargetDataAdapter from '../../scripts/browse-card/browse-cards-target-data-adapter.js';
 import { hideTooltipOnScroll } from '../../scripts/browse-card/browse-card-tooltip.js';
+import { setCoveoCountAsBlockAttribute, setTargetDataAsBlockAttribute } from '../../scripts/analytics/analytics.js';
 
 let placeholders = {};
 try {
@@ -242,37 +243,6 @@ export function updateCopyFromTarget(data, heading, subheading, taglineCta, tagl
       taglineParentBlock?.remove();
     }
   }
-}
-
-/**
- * Sets target data as a data attribute on the given block element.
- *
- * This function checks if the provided `data` object contains a `meta` property.
- * If the `meta` property exists, it serializes the metadata as a JSON string and
- * adds it to the specified block element as a custom data attribute `data-analytics-target-meta`.
- *
- * @param {Object} data - The data returned from target.
- * @param {HTMLElement} block - The DOM element to which the meta data will be added as an attribute.
- *
- */
-export function setTargetDataAsBlockAttribute(data, block) {
-  if (data?.meta) {
-    block.setAttribute('data-analytics-target-meta', JSON.stringify(data?.meta));
-  }
-}
-
-/**
- * Adds a data-analytics-coveo-meta attribute to each recommended-content block on the page.
- * Value is in the format block-name-coveo-X, where X represents the order of the block on the page.
- */
-function setCoveoCountAsBlockAttribute() {
-  const recommendedBlocks = document.querySelectorAll('.recommended-content.block');
-  let coveoCount = 1;
-
-  recommendedBlocks.forEach((block) => {
-    block.setAttribute('data-analytics-coveo-meta', `recommended-content-coveo-${coveoCount}`);
-    coveoCount += 1;
-  });
 }
 
 // fetch list of all interests
@@ -564,7 +534,7 @@ export default async function decorate(block) {
       if (coveoFlowDetection) {
         headerContainer.innerHTML = headingElement.innerHTML;
         descriptionContainer.innerHTML = descriptionElement.innerHTML;
-        setCoveoCountAsBlockAttribute();
+        setCoveoCountAsBlockAttribute(block);
         block.style.display = 'block';
       }
